replace edited geff file with a gds theme set
a quick fix for now.

aim is to strip out custom styling from geff and replace it with a custom GDS theme layer that can be applied over geff. At the moment I have pulled out the styles but there are some things to fix especially layout.

Next steps for this would be to tidy up the GDS theme files and move them back into geff.